Creating skeleton of Leaderboard page
