+++ src/app/components/MobileMenu.tsx
@@ -9,8 +9,9 @@ import {
   Users,
   Trophy,
   Menu,
+  Clock
 } from "lucide-react";
-
+import { IconBallFootball } from "@tabler/icons-react";
 import { Button } from "@/components/ui/button";
 import {
   Sheet,
@@ -70,6 +71,14 @@ export default function MobileMenu() {
               onClick={() => handleNavigation('/games')}
             >
               <Trophy className="h-[1.05rem] w-[1.05rem]" />
+              <span>Leaderboard</span>
+            </Button>
+          <Button
+              variant="ghost"
+              className={getButtonClass('/games')}
+              onClick={() => handleNavigation('/games')}
+            >
+              <IconBallFootball className="h-[1.05rem] w-[1.05rem]" />
               <span>Games</span>
             </Button>
             <Button
