+++ src/app/leaderboard/page.tsx
@@ -1,0 +1,4 @@
+import LeaderboardStats from '@/app/components/LeaderboardStats';
+
+// In your leaderboard page component:
+<LeaderboardStats />
